#127: Moved src/actions_test.rs to hate/examples/actions.rs
