+++ src/screen/main_menu.rs
@@ -8,7 +8,6 @@ enum Command {
     Exit,
     Start,
     GuiTest,
-    ActionsTest,
 }
 
 #[derive(Debug)]
@@ -24,12 +23,9 @@ impl MainMenu {
             let sprite_exit = gui::text_sprite(context, "exit", 0.1);
             let sprite_start = gui::text_sprite(context, "start", 0.1);
             let sprite_gui_test = gui::text_sprite(context, "gui test", 0.1);
-            let sprite_actions_test = gui::text_sprite(context, "actions test", 0.1);
             let button_id_exit = gui.add_button(context, sprite_exit, Command::Exit);
             let button_id_start = gui.add_button(context, sprite_start, Command::Start);
             let button_id_gui_test = gui.add_button(context, sprite_gui_test, Command::GuiTest);
-            let button_id_actions_test =
-                gui.add_button(context, sprite_actions_test, Command::ActionsTest);
             let anchor = gui::Anchor {
                 vertical: gui::VAnchor::Middle,
                 horizontal: gui::HAnchor::Middle,
@@ -40,7 +36,6 @@ impl MainMenu {
                 direction,
                 vec![
                     button_id_exit,
-                    button_id_actions_test,
                     button_id_gui_test,
                     button_id_start,
                 ],
@@ -59,11 +54,6 @@ impl MainMenu {
         context.add_command(hate::screen::Command::Push(screen));
     }
 
-    fn open_screen_actions_test(&mut self, context: &mut Context) {
-        let screen = Box::new(screen::ActionsTest::new(context));
-        context.add_command(hate::screen::Command::Push(screen));
-    }
-
     fn start_new_game(&mut self, context: &mut Context) {
         let game_screen = Box::new(screen::Game::new(context));
         context.add_command(hate::screen::Command::Push(game_screen));
@@ -78,7 +68,6 @@ impl MainMenu {
         while let Some(command) = self.gui.try_recv() {
             match command {
                 Command::GuiTest => self.open_screen_gui_test(context),
-                Command::ActionsTest => self.open_screen_actions_test(context),
                 Command::Start => self.start_new_game(context),
                 Command::Exit => self.exit(context),
             }
