+++ src/screen/mod.rs
@@ -1,9 +1,7 @@
 mod main_menu;
 mod game;
 mod gui_test;
-mod actions_test;
 
 pub use self::main_menu::MainMenu;
 pub use self::game::Game;
 pub use self::gui_test::GuiTest;
-pub use self::actions_test::ActionsTest;
